Full response to warehouse

+++ src/main/java/ba/unsa/etf/si/mainserver/responses/warehouse/ProductQuantityResponse.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 public class ProductQuantityResponse {
     private ProductResponse product;
     private Long quantity;
-    private Long availableQuantity;
+    private double availableQuantity;
 }
